Main menu, buttons interections, bug fixed
arranged build scene position.

arranged glitches such as score clashing with the - swymbol.

implemented a button to take me to menu scene from win scene.

implemented buttons in main menu and and assigned the onclick functions on the start and exit buttons.

increased the speed of the ball in all game levels it was too slow.